changed tests structure, added tests for habit serializer
