+++ habits/tests/test_views.py
@@ -1,5 +1,3 @@
-from django.core.management import call_command
-from django.test import TestCase
 from django.urls import reverse
 from rest_framework.test import APITestCase, APIClient
 from rest_framework import status
@@ -7,28 +5,6 @@ from rest_framework import status
 from django.utils import timezone
 from habits.models import Habit
 from users.models import User
-
-
-class HabitModelTest(TestCase):
-    """Тесты для модели Habit"""
-
-    def setUp(self):
-        self.user = User.objects.create_user(username='testuser', email='testuser@test.ru', password='testpass')
-
-        Habit.objects.create(
-            created_by=self.user,
-            place='Home',
-            time_required=30,
-            time=timezone.now().time(),
-            action='Do something',
-            is_pleasant=False
-        )
-
-    def test_create_habit(self):
-        habit = Habit.objects.get(id=1)
-        self.assertEqual(habit.action, 'Do something')
-        self.assertEqual(habit.created_by, self.user)
-        self.assertFalse(habit.is_pleasant)
 
 
 class HabitsListTest(APITestCase):
@@ -41,7 +17,6 @@ class HabitsListTest(APITestCase):
         )
         self.client.force_authenticate(user=self.user)
 
-        # Создание тестовых привычек
         self.habit1 = Habit.objects.create(
             place='Place1',
             time_required=30,
@@ -104,7 +79,6 @@ class HabitsUpdateTest(APITestCase):
         )
         self.client.force_authenticate(user=self.user)
 
-        # Создание тестовых привычек
         self.habit1 = Habit.objects.create(
             place='Place1',
             time_required=30,
@@ -137,7 +111,6 @@ class HabitsDeleteTest(APITestCase):
         )
         self.client.force_authenticate(user=self.user)
 
-        # Создание тестовых привычек
         self.habit1 = Habit.objects.create(
             place='Place1',
             time_required=30,
@@ -165,7 +138,6 @@ class PublicHabitsListTest(APITestCase):
         )
         self.client.force_authenticate(user=self.user)
 
-        # Создание тестовых привычек
         self.habit1 = Habit.objects.create(
             place='Place1',
             time_required=30,
